chore: updated constants
