+++ app/container-imp.ts
@@ -55,6 +55,8 @@ import PersonCredentialLoading from './src/bcwallet-theme/features/person-flow/s
 import { pages } from './src/components/OnboardingPages'
 import {
   AttestationRestrictions,
+  DIGITAL_SERVICES_CARD_CREDENTIAL_DEFINITION_IDS,
+  DIGITAL_SERVICES_CARD_SCHEMA_IDS,
   Mode,
   appHelpUrl,
   appleAppStoreUrl,
@@ -135,18 +137,8 @@ export class AppContainer implements Container {
 
     this._container.registerInstance(TOKENS.CRED_HELP_ACTION_OVERRIDES, [
       {
-        credDefIds: [
-          'RGjWbW1eycP7FrMf4QJvX8:3:CL:13:Person',
-          'KCxVC8GkKywjhWJnUfCmkW:3:CL:20:PersonQA',
-          '7xjfawcnyTUcduWVysLww5:3:CL:28075:PersonSIT',
-          'XpgeQa93eZvGSZBZef3PHn:3:CL:28075:PersonDEV',
-        ],
-        schemaIds: [
-          'RGjWbW1eycP7FrMf4QJvX8:2:Person:1.0',
-          'KCxVC8GkKywjhWJnUfCmkW:2:Person:1.0',
-          '7xjfawcnyTUcduWVysLww5:2:Person:1.0',
-          'XpgeQa93eZvGSZBZef3PHn:2:Person:1.0',
-        ],
+        credDefIds: DIGITAL_SERVICES_CARD_CREDENTIAL_DEFINITION_IDS,
+        schemaIds: DIGITAL_SERVICES_CARD_SCHEMA_IDS,
         action: (navigation: NavigationProp<ReactNavigation.RootParamList>) => {
           navigation.getParent()?.navigate(Stacks.NotificationStack, {
             screen: Screens.CustomNotification,
@@ -274,8 +266,8 @@ export class AppContainer implements Container {
           headerLeft: createMainSettingsHeaderButton(),
         },
       })
-      this._container.registerInstance(TOKENS.COMPONENT_CRED_SUBHEADER, CredentialDetailsSubHeader)
     }
+    this._container.registerInstance(TOKENS.COMPONENT_CRED_SUBHEADER, CredentialDetailsSubHeader)
     this._container.registerInstance(TOKENS.COMPONENT_RECORD, Record)
     this._container.registerInstance(TOKENS.CACHE_CRED_DEFS, [
       // { did: "4WW6792ksq62UroZyfd6nQ", id: "4WW6792ksq62UroZyfd6nQ:3:CL:1098:SellingItRight" },
